Moved crackers to Crypt.py Renamed cracker.py to functions.py
